Dirty WIP hack on Web Component wrappers
Currently this just globs all the components in src/ui/components/
and writes them to a file statically that can be imported at runtime.

Eventually we should figure out what the API is for opting in/out of
WC wrappers and trim it to that list. E.g. we could parse each component
looking for the `@WebComponent` decorator or something.