Update ember-cli-blueprint-test-helpers to latest.
Prior to this version, running `npm test && npm test` would result
in a very cryptic error RE: recursive directory structure passed to
`walkSync`. This was happening because (for testing purposes)
`ember-cli-blueprint-test-helpers` "linked" the addon undertest
so that it could test the generators.  Unfortunatley, this means adding
a symlink in `node_modules/<project name>` pointing to `../` (project root).

This update correctly cleans up after itself, and removes the symlink
created to prevent these sorts of issues...