Introduce a public hook for packaging
This introduces a `package` hook that can be overriden by the consuming
application. We retain the current build pipeline with rollup for the
time being, but this gives us an extensible way of doing a lot of
expirementation.

```
public package(jsTree: Tree, cssTree: Tree, publicTree: Tree, htmlTree:
Tree): Tree;
```

This PR also introduces the addon registry so we can utilizes existing
ember-cli preprocessing plugins.