Remove `maybePerformDeprecatedCssConcat`.
This was added as a result of a review I did where I observed that this
behavior was being lost during the styles pipeline refactor. Unfortunatley,
I do not think that we can actually manage to pull off this deprecation
(and I do not think it is a very likely / common issue to have either).

I am removing this because there is no way to appease the deprecation
if you legitimately want/need multiple CSS files (e.g. you want to lazily
load one or something).