Move timeout to avoid ordering issue during tests.
The timeout was previously done in `toTree` because that module
was the first one that had to kick off all the requires and whatnot
for `babel-preset-env`.

This moves the timeout up to apply to all of the describe blocks, so
that we don't have to worry about the specific ordering.